remove unnecessary casts when printing object addresses
casting pointers to 'long' causes warnings in 64-bit Windows builds (because 'long' is really 32-bit)

Note that this cast only appears in these three spots.
There are literally hundreds of places where we *don't* cast the pointer before passing it to ".x%lx"

+++ src/g_graph.c
@@ -747,7 +747,7 @@ static void graph_vis(t_gobj *gr, t_glist *parent_glist, int vis)
             i -= glist_fontheight(x);
             sys_vgui(".x%lx.c create text %d %d -text {%s} -anchor nw "
                 "-font {{%s} -%d %s} -tags [list %s label graph]\n",
-                (long)glist_getcanvas(x),  x1, i,
+                glist_getcanvas(x),  x1, i,
                 arrayname->s_name, sys_font,
                 fs, sys_fontweight, tag);
         }
